botCeibe/requirements.txt
ccxt>=4.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.0.6

flask>=3.0.0
